requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
assemblyai==0.33.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
mutagen==1.47.0
numpy==2.1.0
opencv-python==4.10.0.84
pafy==0.5.5
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyJWT==2.9.0
pyparsing==3.1.4
pytube==15.0.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SpeechRecognition==3.10.4
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==13.0.1
youtube-dl==2021.12.17
yt-dlp==2024.8.6
